+++ libs/vre/ui/nested-menu/.eslintrc.json
@@ -1,0 +1,11 @@
+{
+  "extends": ["../../../../.eslintrc-angular.json"],
+  "overrides": [
+    {
+      "files": ["*.ts"],
+      "rules": {
+        "@angular-eslint/prefer-standalone": "off"
+      }
+    }
+  ]
+}
