+++ libs/vre/ui/nested-menu/README.md
@@ -1,0 +1,7 @@
+# vre-ui-nested-menu
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test vre-ui-nested-menu` to execute the unit tests.
