+++ libs/vre/ui/nested-menu/jest.config.ts
@@ -1,0 +1,8 @@
+/* eslint-disable */
+export default {
+    displayName: 'vre-ui-nested-menu',
+    preset: '../../../../jest.preset.js',
+    coverageDirectory:
+        '../../../../coverage/libs/vre/ui/nested-menu',
+  collectCoverageFrom: ['src/**/*.ts', '!src/**/*.spec.ts', '!src/**/*.d.ts', '!src/test-setup.ts', '!src/index.ts'],
+};
