+++ libs/vre/ui/nested-menu/project.json
@@ -1,0 +1,28 @@
+{
+  "name": "vre-ui-nested-menu",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/vre/ui/nested-menu/src",
+  "prefix": "dasch-swiss",
+  "projectType": "library",
+  "tags": ["type:ui"],
+  "targets": {
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/vre/ui/nested-menu/jest.config.ts",
+        "tsConfig": "libs/vre/ui/nested-menu/tsconfig.spec.json"
+      }
+    },
+    "lint": {
+      "executor": "@nx/eslint:lint",
+      "outputs": ["{options.outputFile}"]
+    },
+    "tsc-errors": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "tsc --noEmit --project {projectRoot}/tsconfig.lib.json"
+      }
+    }
+  }
+}
