+++ tsconfig.base.json
@@ -42,6 +42,7 @@
       "@dasch-swiss/vre/shared/assets/status-msg": ["libs/vre/shared/assets/status-msg/src/index.ts"],
       "@dasch-swiss/vre/shared/calendar": ["libs/vre/shared/calendar/src/index.ts"],
       "@dasch-swiss/vre/ui/date-picker": ["libs/vre/ui/date-picker/src/index.ts"],
+      "@dasch-swiss/vre/ui/nested-menu": ["libs/vre/ui/nested-menu/src/index.ts"],
       "@dasch-swiss/vre/ui/notification": ["libs/vre/ui/notification/src/index.ts"],
       "@dasch-swiss/vre/ui/progress-indicator": ["libs/vre/ui/progress-indicator/src/index.ts"],
       "@dasch-swiss/vre/ui/string-literal": ["libs/vre/ui/string-literal/src/index.ts"],
